+++ src/main/java/com/telepathicgrunt/tact/TACT.java
@@ -13,7 +13,6 @@ import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.PackSource;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
@@ -21,16 +20,20 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.AddPackFindersEvent;
 import net.minecraftforge.event.entity.living.MobEffectEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.ModContainer;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.config.ConfigTracker;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.loading.FMLPaths;
 import net.minecraftforge.resource.PathPackResources;
-import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
+import java.lang.reflect.Method;
+import java.nio.file.Path;
 import java.util.Optional;
 
 @Mod(TACT.MODID)
@@ -77,20 +80,63 @@ public class TACT {
                         .findResource("datapacks/apply_tag_adjustments");
 
                 var pack = Pack.readMetaAndCreate(
-                        "builtin/apply_tag_adjustments",
+                        "tact/apply_tag_adjustments",
                         Component.literal("TACT - Adjusted Tags"),
-                        false,
+                        true,
                         (path) -> new PathPackResources(path, false, resourcePath),
                         PackType.SERVER_DATA,
                         Pack.Position.BOTTOM,
-                        PackSource.BUILT_IN);
+                        PackSource.DEFAULT);
 
                 event.addRepositorySource((packConsumer) -> packConsumer.accept(pack));
             }
+
+
+            if (event.getPackType() == PackType.CLIENT_RESOURCES) {
+                loadTACTConfigsEarly();
+
+                if (Config.restoreEndStory) {
+                    var resourcePath =  ModList.get().getModFileById(MODID).getFile()
+                            .findResource("resourcepacks/restore_end_story");
+
+                    var pack = Pack.readMetaAndCreate(
+                            "tact/restore_end_story",
+                            Component.literal("TACT - Restore End Story"),
+                            true,
+                            (path) -> new PathPackResources(path, false, resourcePath),
+                            PackType.CLIENT_RESOURCES,
+                            Pack.Position.TOP,
+                            PackSource.DEFAULT);
+
+                    event.addRepositorySource((packConsumer) -> packConsumer.accept(pack));
+                }
+            }
+
         }
         catch(Exception ex) {
             throw new RuntimeException(ex);
         }
+    }
+
+    private static void loadTACTConfigsEarly() {
+        Optional<? extends ModContainer> modContainerById = ModList.get().getModContainerById(MODID);
+        modContainerById.ifPresent(container -> {
+            ConfigTracker.INSTANCE.configSets()
+                .get(ModConfig.Type.COMMON)
+                .forEach(c -> {
+                    if (c.getFileName().equals("tact-common.toml")) {
+                        try {
+                            Method method = ConfigTracker.INSTANCE.getClass().getDeclaredMethod("openConfig", ModConfig.class, Path.class);
+                            method.setAccessible(true);
+                            method.invoke(ConfigTracker.INSTANCE, c, FMLPaths.CONFIGDIR.get());
+                            method.setAccessible(false);
+                        }
+                        catch (Exception e) {
+                            throw new RuntimeException(e);
+                        }
+                    }
+                });
+        });
     }
 
     private void stunEffectAdjustment(final MobEffectEvent.Added event) {