+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### **(V.1.0.1 Changes) (1.20.1 Minecraft)**
 
+#### End Story:
+Added config to revert End Story back to normal when beating Minecraft. This undoes the story replacement that Alex's Caves does.
+ This config is set to false by default but if set to true, the story is put back to normal.
+
 #### Tags:
 Removed and adjusted some tags as Alex's Caves mod now has updated to include many tags.