+++ CHANGELOG.md
@@ -3,5 +3,8 @@
 #### Entities:
 Added config options to set every Alex's Caves entity's health and attack damage.
 
+#### Blocks:
+Added Bass note for Note Block behavior to Pewen and Thornwood log and planks. This makes Every Compat mod now work to create Pewen and Thornwood variant blocks.
+
 #### Tags:
 Added Alex's Caves saplings to the `minecraft:saplings` item tag.